move `fritz` api initi to coordinator setup
