+++ homeassistant/components/fritzbox/__init__.py
@@ -4,49 +4,23 @@ from __future__ import annotations
 
 from abc import ABC, abstractmethod
 
-from pyfritzhome import Fritzhome, FritzhomeDevice, LoginError
+from pyfritzhome import FritzhomeDevice
 from pyfritzhome.devicetypes.fritzhomeentitybase import FritzhomeEntityBase
-from requests.exceptions import ConnectionError as RequestConnectionError
 
 from homeassistant.components.binary_sensor import DOMAIN as BINARY_SENSOR_DOMAIN
-from homeassistant.const import (
-    CONF_HOST,
-    CONF_PASSWORD,
-    CONF_USERNAME,
-    EVENT_HOMEASSISTANT_STOP,
-    UnitOfTemperature,
-)
+from homeassistant.const import EVENT_HOMEASSISTANT_STOP, UnitOfTemperature
 from homeassistant.core import Event, HomeAssistant
-from homeassistant.exceptions import ConfigEntryAuthFailed, ConfigEntryNotReady
 from homeassistant.helpers.device_registry import DeviceEntry, DeviceInfo
 from homeassistant.helpers.entity import EntityDescription
 from homeassistant.helpers.entity_registry import RegistryEntry, async_migrate_entries
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
 from .const import DOMAIN, LOGGER, PLATFORMS
-from .coordinator import FritzboxDataUpdateCoordinator
-from .model import FritzboxConfigEntry
+from .coordinator import FritzboxConfigEntry, FritzboxDataUpdateCoordinator
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: FritzboxConfigEntry) -> bool:
     """Set up the AVM FRITZ!SmartHome platforms."""
-    fritz = Fritzhome(
-        host=entry.data[CONF_HOST],
-        user=entry.data[CONF_USERNAME],
-        password=entry.data[CONF_PASSWORD],
-    )
-
-    try:
-        await hass.async_add_executor_job(fritz.login)
-    except RequestConnectionError as err:
-        raise ConfigEntryNotReady from err
-    except LoginError as err:
-        raise ConfigEntryAuthFailed from err
-
-    entry.runtime_data.fritz = fritz
-
-    has_templates = await hass.async_add_executor_job(fritz.has_templates)
-    LOGGER.debug("enable smarthome templates: %s", has_templates)
 
     def _update_unique_id(entry: RegistryEntry) -> dict[str, str] | None:
         """Update unique ID of entity entry."""
@@ -70,16 +44,16 @@ async def async_setup_entry(hass: HomeAssistant, entry: FritzboxConfigEntry) -> 
 
     await async_migrate_entries(hass, entry.entry_id, _update_unique_id)
 
-    coordinator = FritzboxDataUpdateCoordinator(hass, entry.entry_id, has_templates)
+    coordinator = FritzboxDataUpdateCoordinator(hass, entry.entry_id)
     await coordinator.async_setup()
 
-    entry.runtime_data.coordinator = coordinator
+    entry.runtime_data = coordinator
 
     await hass.config_entries.async_forward_entry_setups(entry, PLATFORMS)
 
     def logout_fritzbox(event: Event) -> None:
         """Close connections to this fritzbox."""
-        fritz.logout()
+        coordinator.fritz.logout()
 
     entry.async_on_unload(
         hass.bus.async_listen_once(EVENT_HOMEASSISTANT_STOP, logout_fritzbox)
@@ -99,7 +73,7 @@ async def async_remove_config_entry_device(
     hass: HomeAssistant, entry: FritzboxConfigEntry, device: DeviceEntry
 ) -> bool:
     """Remove Fritzbox config entry from a device."""
-    coordinator = entry.runtime_data.coordinator
+    coordinator = entry.runtime_data
 
     for identifier in device.identifiers:
         if identifier[0] == DOMAIN and (
