+++ homeassistant/components/fritzbox/diagnostics.py
@@ -8,7 +8,7 @@ from homeassistant.components.diagnostics import async_redact_data
 from homeassistant.const import CONF_PASSWORD, CONF_USERNAME
 from homeassistant.core import HomeAssistant
 
-from .model import FritzboxConfigEntry
+from .coordinator import FritzboxConfigEntry
 
 TO_REDACT = {CONF_USERNAME, CONF_PASSWORD}
 
@@ -17,7 +17,7 @@ async def async_get_config_entry_diagnostics(
     hass: HomeAssistant, entry: FritzboxConfigEntry
 ) -> dict[str, Any]:
     """Return diagnostics for a config entry."""
-    coordinator = entry.runtime_data.coordinator
+    coordinator = entry.runtime_data
 
     diag_data = {
         "entry": async_redact_data(entry.as_dict(), TO_REDACT),
