+++ homeassistant/components/fritzbox/model.py
@@ -4,16 +4,9 @@ from __future__ import annotations
 
 from collections.abc import Callable
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, TypedDict
+from typing import TypedDict
 
-from pyfritzhome import Fritzhome, FritzhomeDevice
-
-from homeassistant.config_entries import ConfigEntry
-
-if TYPE_CHECKING:
-    from .coordinator import FritzboxDataUpdateCoordinator
-
-FritzboxConfigEntry = ConfigEntry["FritzboxRuntimeData"]
+from pyfritzhome import FritzhomeDevice
 
 
 class ClimateExtraAttributes(TypedDict, total=False):
@@ -31,11 +24,3 @@ class FritzEntityDescriptionMixinBase:
     """Bases description mixin for Fritz!Smarthome entities."""
 
     suitable: Callable[[FritzhomeDevice], bool]
-
-
-@dataclass
-class FritzboxRuntimeData:
-    """Fritzbox runtime data class."""
-
-    coordinator: FritzboxDataUpdateCoordinator
-    fritz: Fritzhome
