+++ homeassistant/components/fritzbox/sensor.py
@@ -31,7 +31,8 @@ from homeassistant.helpers.typing import StateType
 from homeassistant.util.dt import utc_from_timestamp
 
 from . import FritzBoxDeviceEntity
-from .model import FritzboxConfigEntry, FritzEntityDescriptionMixinBase
+from .coordinator import FritzboxConfigEntry
+from .model import FritzEntityDescriptionMixinBase
 
 
 @dataclass(frozen=True)
@@ -213,7 +214,7 @@ async def async_setup_entry(
     async_add_entities: AddEntitiesCallback,
 ) -> None:
     """Set up the FRITZ!SmartHome sensor from ConfigEntry."""
-    coordinator = entry.runtime_data.coordinator
+    coordinator = entry.runtime_data
 
     @callback
     def _add_entities(devices: set[str] | None = None) -> None:
