+++ tests/components/fritzbox/conftest.py
@@ -9,7 +9,7 @@ import pytest
 def fritz_fixture() -> Mock:
     """Patch libraries."""
     with (
-        patch("homeassistant.components.fritzbox.Fritzhome") as fritz,
+        patch("homeassistant.components.fritzbox.coordinator.Fritzhome") as fritz,
         patch("homeassistant.components.fritzbox.config_flow.Fritzhome"),
     ):
         fritz.return_value.get_prefixed_host.return_value = "http://1.2.3.4"
